Add files via upload
To run this app locally, follow the instructions below:

Download the code from the repository and extract it to a directory of your choice.

Open a terminal and navigate to the project directory.

Run npm install to install the required dependencies.

After installation, run npm start to start the app.

Once the app is running, open a web browser and navigate to http://localhost:3000/.

Enter a list of names separated by commas into the text field and click the "Submit" button to see the names displayed in a table below.

To edit or delete a name, click on the "Edit" or "Delete" button next to the corresponding name in the table.

To stop the app, simply stop the running process in the terminal by pressing Ctrl + C.

That's it! You should now be able to run the app locally on your machine. If you encounter any issues or have any questions, feel free to reach out to the developer (Me!).